pyproject.toml
[project]
name = "faq-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]
